+++ NEWS.md
@@ -4,6 +4,7 @@
 
 ## Updates and Fixes
 
++ `eia2tidy()` amended for quarterly data.
 + Morningstar tickers for `getPrice()` and `getPrices()` functions.
 
 ## Removed