+++ NEWS.md
@@ -1,3 +1,11 @@
+# RTL 0.1.6
+
+## New
+
+## Updates and Fixes
+
+## Removed
+
 # RTL 0.1.5
 
 ## New